get data inside members_modal
